cleanups

+++ macadam/details/math/mc_exp.h
@@ -33,7 +33,7 @@ static MC_TARGET_INLINE double mc_exp(double x)
 #	endif
 }
 
-static MC_TARGET_INLINE double mc_expl(double x)
+static MC_TARGET_INLINE long double mc_expl(long double x)
 {
 #	if MC_TARGET_CPP98
 	return ::expl(x);
