+++ macadam/details/math/mc_fabs.h
@@ -33,7 +33,7 @@ static MC_TARGET_INLINE double mc_fabs(double x)
 #	endif
 }
 
-static MC_TARGET_INLINE double mc_fabsl(double x)
+static MC_TARGET_INLINE long double mc_fabsl(long double x)
 {
 #	if MC_TARGET_CPP98
 	return ::fabsl(x);
