+++ macadam/details/math/mc_floor.h
@@ -33,7 +33,7 @@ static MC_TARGET_INLINE double mc_floor(double x)
 #	endif
 }
 
-static MC_TARGET_INLINE double mc_floorl(double x)
+static MC_TARGET_INLINE long double mc_floorl(long double x)
 {
 #	if MC_TARGET_CPP98
 	return ::floorl(x);
