+++ macadam/details/math/mc_log.h
@@ -33,7 +33,7 @@ static MC_TARGET_INLINE double mc_log(double x)
 #	endif
 }
 
-static MC_TARGET_INLINE double mc_logl(double x)
+static MC_TARGET_INLINE long double mc_logl(long double x)
 {
 #	if MC_TARGET_CPP98
 	return ::logl(x);
